refactor(platform): replace DenoAdapter with SupabasePlatformAdapter and update worker setup
- Replaced DenoAdapter with SupabasePlatformAdapter for enhanced platform-specific functionality
- Added methods to spawn new edge functions and extend their lifetime using EdgeRuntime
- Updated startWorker to manage resources and handle shutdown signals properly
- Modified context creation to include SQL and Supabase clients for message handling
- Export index now consolidates platform adapters, removing legacy DenoAdapter references
- Overall improvements enable better resource management and platform integration in edge
environment
