+++ pkgs/edge-worker/src/_internal/platform.ts
@@ -1,6 +1,6 @@
 export { createAdapter } from '../platform/createAdapter.js';
 export { createLoggingFactory } from '../platform/logging.js';
-export { DenoAdapter } from '../platform/DenoAdapter.js';
+export { SupabasePlatformAdapter } from '../platform/SupabasePlatformAdapter.js';
 export type * from '../platform/types.js';
 // Also re-export everything from platform/index.js
 export * from '../platform/index.js';
